extract price history downloader
