+++ lib/schwab_rb.rb
@@ -22,6 +22,7 @@ require_relative "schwab_rb/option"
 require_relative "schwab_rb/orders/instruments"
 require_relative "schwab_rb/market_hours"
 require_relative "schwab_rb/price_history"
+require_relative "schwab_rb/price_history/downloader"
 require_relative "schwab_rb/movers"
 require_relative "schwab_rb/orders/builder"
 require_relative "schwab_rb/orders/session"
